requirements.txt
dlib==19.18.0
falcon==2.0.0
Flask==1.0.2
matplotlib==3.0.3
numpy==1.17.2
opencv-python==4.1.0.25
Pillow>=6.2.0
tensorflow==1.13.1
keras==2.2.4
scikit-fuzzy>=0.1